town creation

+++ Assets/DialogInfo.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class DialogInfo
 {
-    public long id;
+    public int id;
     public string title;
     public string message;
     public string inputProperty;
     public DialogType type;
 
-    public DialogInfo(long id, string title, string message, string inputProperty, DialogType type)
+    public DialogInfo(int id, string title, string message, string inputProperty, DialogType type)
     {
         this.id = id;
         this.title = title;
